+++ tests/e2e/game/game-options.test.js
@@ -65,6 +65,7 @@ describe("K - Game options", () => {
                     roles: {
                         areRevealedOnDeath: false,
                         sheriff: { hasDoubledVote: false },
+                        lovers: { doRevealRoleToEachOther: true },
                         seer: { isTalkative: false, canSeeRoles: false },
                         guard: { canProtectTwice: true },
                         twoSisters: { wakingUpInterval: 1 },
@@ -86,6 +87,7 @@ describe("K - Game options", () => {
                 expect(game.options.repartition.isHidden).to.be.true;
                 expect(game.options.roles.areRevealedOnDeath).to.be.false;
                 expect(game.options.roles.sheriff.hasDoubledVote).to.be.false;
+                expect(game.options.roles.lovers.doRevealRoleToEachOther).to.be.true;
                 expect(game.options.roles.seer.isTalkative).to.be.false;
                 expect(game.options.roles.seer.canSeeRoles).to.be.false;
                 expect(game.options.roles.cupid.mustWinWithLovers).to.be.false;